moved out store into a separate file, refactored action methods to use unified api
